+++ ThisIsTheWayshrine/ThisIsTheWayshrineSettings.lua
@@ -24,7 +24,7 @@ function TITW.BuildMenu()
       TITW.SV.enableJumping = var
       if var then
         TITW.checkGuildMembersCurrentZoneAndJump()
-        EVENT_MANAGER:RegisterForUpdate("TITW_CheckAndJump", 30000, TITW.checkGuildMembersCurrentZoneAndJump)
+        EVENT_MANAGER:RegisterForUpdate("TITW_CheckAndJump", 15000, TITW.checkGuildMembersCurrentZoneAndJump)
       else
         EVENT_MANAGER:UnregisterForUpdate("TITW_CheckAndJump")
       end