+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## 2026.01.18
 - Increase frequency of jumping and failsafe in case of canceling travel
 - Add pull request template
+- Fix first time use
 
 ## 2026.01.17
 - Minimum features working