+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 - Fix first time use
 - Clean up addon load
 - Attempt delaying the initial jump on first time use
+- Use register update
 
 ## 2026.01.17
 - Minimum features working