+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - Clean up addon load
 - Attempt delaying the initial jump on first time use
 - Use register update
+- Tune register update frequency
 
 ## 2026.01.17
 - Minimum features working