+++ CHANGELOG.md
@@ -8,7 +8,7 @@
 - Tune register update frequency
 - Use a hybrid update and recursive approach
 - Move call later checkAndJump out of conditional
-- Attempt to fix recursion
+- Introduce stalled check instead of calling recursively
 
 ## 2026.01.17
 - Minimum features working