+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 - Use a hybrid update and recursive approach
 - Move call later checkAndJump out of conditional
 - Introduce stalled check instead of calling recursively
+- Increase call later duration
 
 ## 2026.01.17
 - Minimum features working