+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## 2026.01.19
 - Decrease stalled check counter
 - Address case of no valid jumps available
+- Fix some excessive calling itself
 
 ## 2026.01.18
 - Increase frequency of jumping and failsafe in case of canceling travel